+++ validator/src/cli.rs
@@ -18,7 +18,6 @@ use {
         hidden_unless_forced,
         input_validators::{
             is_parsable, is_pubkey, is_pubkey_or_keypair, is_slot, is_url_or_moniker,
-            validate_cpu_ranges,
         },
     },
     solana_clock::Slot,
@@ -190,41 +189,6 @@ fn deprecated_arguments() -> Vec<DeprecatedArg> {
             .validator(is_parsable::<usize>)
             .help("Specify which CPU core PoH is pinned to. Use --poh-pinned-cpu-core instead"),
         replaced_by: "poh-pinned-cpu-core",
-    );
-    add_arg!(
-        // deprecated in v4.1.0
-        Arg::with_name("experimental_retransmit_xdp_cpu_cores")
-            .long("experimental-retransmit-xdp-cpu-cores")
-            .takes_value(true)
-            .value_name("CPU_LIST")
-            .conflicts_with("xdp_cpu_cores")
-            .conflicts_with("no_xdp")
-            .validator(|value| {
-                validate_cpu_ranges(value, "--experimental-retransmit-xdp-cpu-cores")
-            })
-            .help("CPU cores to reserve for XDP. Use --xdp-cpu-cores instead"),
-        replaced_by: "xdp-cpu-cores",
-    );
-    add_arg!(
-        // deprecated in v4.1.0
-        Arg::with_name("experimental_retransmit_xdp_interface")
-            .long("experimental-retransmit-xdp-interface")
-            .takes_value(true)
-            .value_name("INTERFACE")
-            .conflicts_with("xdp_interface")
-            .conflicts_with("no_xdp")
-            .help("Network interface to use for XDP. Use --xdp-interface instead"),
-        replaced_by: "xdp-interface",
-    );
-    add_arg!(
-        // deprecated in v4.1.0
-        Arg::with_name("experimental_retransmit_xdp_zero_copy")
-            .long("experimental-retransmit-xdp-zero-copy")
-            .takes_value(false)
-            .conflicts_with("xdp_zero_copy")
-            .conflicts_with("no_xdp")
-            .help("Enable XDP zero copy. Use --xdp-zero-copy instead"),
-        replaced_by: "xdp-zero-copy",
     );
     add_arg!(
         // deprecated in v4.3.0